backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
supabase==2.5.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
pydantic==2.7.1
pandas==2.2.2
numpy==1.26.4
yfinance==0.2.40
requests==2.32.3
python-multipart==0.0.9
beautifulsoup4==4.12.3
